# api/openapi/rest-api.yaml
openapi: 3.0.3
info:
  title: Maestro Management API
  description: REST API for managing clusters, agents, tasks, and executions
  version: 1.0.0
  contact:
    name: Maestro Team
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/clusters:
    get:
      tags: [Clusters]
      summary: List all clusters
      operationId: listClusters
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of clusters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Cluster'
    post:
      tags: [Clusters]
      summary: Create a new cluster
      operationId: createCluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClusterRequest'
      responses:
        '201':
          description: Cluster created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cluster'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/clusters/{id}:
    get:
      tags: [Clusters]
      summary: Get cluster by ID
      operationId: getCluster
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cluster'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Clusters]
      summary: Delete cluster
      operationId: deleteCluster
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Cluster deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/clusters/{clusterId}/agents:
    get:
      tags: [Agents]
      summary: List agents in a cluster
      operationId: listAgentsByCluster
      parameters:
        - name: clusterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Agent'
  /api/v1/agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent by ID
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Agents]
      summary: Delete agent
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/agents/{id}/executions:
    get:
      tags: [Agents]
      summary: List executions for an agent
      operationId: listAgentExecutions
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskExecution'
  /api/v1/executions:
    get:
      tags: [Executions]
      summary: List all executions
      operationId: listExecutions
      parameters:
        - name: cluster_id
          in: query
          schema:
            type: string
            format: uuid
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
        - name: task_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed, skipped]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskExecution'
  /api/v1/executions/{id}:
    get:
      tags: [Executions]
      summary: Get execution by ID
      operationId: getExecution
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskExecution'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/debug-tasks:
    post:
      tags: [Debug Tasks]
      summary: Create a debug task
      operationId: createDebugTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDebugTaskRequest'
      responses:
        '201':
          description: Debug task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebugTask'
  /api/v1/debug-tasks/{id}:
    get:
      tags: [Debug Tasks]
      summary: Get debug task by ID
      operationId: getDebugTask
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Debug task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebugTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/agents/{agentId}/debug-tasks:
    get:
      tags: [Debug Tasks]
      summary: List debug tasks for an agent
      operationId: listDebugTasksByAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of debug tasks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/DebugTask'
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks for a cluster
      operationId: listTasks
      parameters:
        - name: cluster_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: include_deleted
          in: query
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/v1/tasks/reorder:
    post:
      tags: [Tasks]
      summary: Reorder tasks
      operationId: reorderTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTasksRequest'
      responses:
        '204':
          description: Tasks reordered
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      tags: [Tasks]
      summary: Update task
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      tags: [Tasks]
      summary: Delete task
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Task deleted
  /api/v1/tasks/{id}/reset-executions:
    post:
      tags: [Tasks]
      summary: Reset task executions
      operationId: resetTaskExecutions
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '204':
          description: Executions reset
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - total
        - limit
        - offset
    Cluster:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - created_at
        - updated_at
    CreateClusterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cluster_id:
          type: string
          format: uuid
        hostname:
          type: string
        status:
          type: string
          enum: [active, inactive]
        last_heartbeat:
          type: string
          format: date-time
        registered_at:
          type: string
          format: date-time
      required:
        - id
        - cluster_id
        - hostname
        - status
        - last_heartbeat
        - registered_at
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cluster_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [exec]
        order:
          type: integer
        blocking:
          type: boolean
        config:
          $ref: '#/components/schemas/TaskConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - cluster_id
        - name
        - type
        - order
        - blocking
        - config
        - created_at
        - updated_at
...
    CreateTaskRequest:
      type: object
      properties:
        cluster_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [exec]
        blocking:
          type: boolean
          default: false
        config:
          $ref: '#/components/schemas/TaskConfig'
      required:
        - cluster_id
        - name
        - type
        - config
...
    ReorderTasksRequest:
      type: object
      properties:
        cluster_id:
          type: string
          format: uuid
        task_ids:
          type: array
          items:
            type: string
            format: uuid
      required:
        - cluster_id
        - task_ids
    TaskExecution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, success, failed, skipped]
        output:
          type: string
        exit_code:
          type: integer
          nullable: true
        error:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - task_id
        - agent_id
        - status
        - started_at
    DebugTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        command:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, skipped]
        output:
          type: string
        exit_code:
          type: integer
          nullable: true
        error:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - agent_id
        - command
        - status
        - created_at
    CreateDebugTaskRequest:
      type: object
      properties:
        agent_id:
          type: string
          format: uuid
        command:
          type: string
      required:
        - agent_id
        - command
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'